Initialized project with tsconfig.json
