+++ packages/app/components/EventCard/EventCard.ts
