+++ packages/app/components/EventCard/EventCardText.ts
