+++ packages/app/components/EventCard/FaceIcon.ts
