+++ packages/app/package.json
@@ -24,17 +24,23 @@
     "expo-web-browser": "~12.0.0",
     "react": "18.1.0",
     "react-native": "0.70.5",
-    "react-native-safe-area-context": "4.4.1",
-    "react-native-screens": "~3.18.0",
     "react-native-gesture-handler": "~2.8.0",
     "react-native-reanimated": "~2.12.0",
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0",
     "react-test-renderer": "^18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
     "@testing-library/react-native": "^11.5.0",
+    "@tsconfig/react-native": "^2.0.3",
+    "@types/jest": "^29.4.0",
+    "@types/react": "^18.0.28",
+    "@types/react-native": "^0.71.2",
+    "@types/react-test-renderer": "^18.0.0",
+    "babel-plugin-module-resolver": "^4.1.0",
     "jest": "^29.3.1",
-    "babel-plugin-module-resolver": "^4.1.0"
+    "typescript": "^4.9.5"
   },
   "private": true
 }
